migrate tabs
